fix delete page
