+++ apps/web/components/ui/UnlinkCategoryModal/UnlinkCategoryModal.vue
@@ -22,7 +22,7 @@
           aria-label="deleteButton"
           data-testid="delete-btn"
           class="bg-red-700 w-full py-2 rounded-md text-white"
-          @click="handleDelete(id)"
+          @click="deletePage(id, pageName)"
         >
           Delete page
         </button>
@@ -43,7 +43,7 @@
 <script setup lang="ts">
 import { SfIconClose } from '@storefront-ui/vue';
 
-const { unlinkModalOpen, toggleDeleteModal, deletePage } = useSiteConfiguration();
+const { unlinkModalOpen, toggleDeleteModal, deletePage } = useCategorySettings();
 const { getCategoryId } = useCategoryIdHelper();
 const { pages } = await usePages();
 const pageName = ref('');
@@ -76,8 +76,5 @@ watch(
 
 const closeModal = () => {
   toggleDeleteModal(false);
-};
-const handleDelete = (id: number) => {
-  deletePage(id, pageName.value);
 };
 </script>
