+++ apps/web/composables/useCategorySettings/types.ts
@@ -1,4 +1,4 @@
-import type { DrawerView } from '~/composables/useSiteConfiguration/types';
+import { DrawerView, UseSiteConfigurationState } from '~/composables/useSiteConfiguration/types';
 import type { Category } from '@plentymarkets/shop-api';
 
 export interface useCategoryConfigurationState {
@@ -9,7 +9,9 @@ export interface useCategoryConfigurationState {
   drawerExtraOpen: boolean;
   drawerView: DrawerView | null;
   initialData: Category;
+  unlinkModalOpen: boolean;
 }
+export type ToggleDeleteModal = (value: boolean) => void;
 
 export interface useCategorySettings {
   data: Ref<useCategoryConfigurationState['data']>;
@@ -19,6 +21,9 @@ export interface useCategorySettings {
   drawerExtraOpen: Readonly<Ref<useCategoryConfigurationState['drawerExtraOpen']>>;
   drawerView: Readonly<Ref<useCategoryConfigurationState['drawerView']>>;
   fetchCategorySettings: (categoryId: number) => Promise<Category | null>;
+  unlinkModalOpen: Readonly<Ref<useCategoryConfigurationState['unlinkModalOpen']>>;
+  deletePage: (pageId: number, pageName: string) => void;
+  toggleDeleteModal: ToggleDeleteModal;
 }
 
 export type useCategorySettingsReturn = () => useCategorySettings;
