+++ apps/web/composables/useCategorySettings/useCategorySettings.ts
@@ -11,7 +11,9 @@ export const useCategorySettings: useCategorySettingsReturn = (settingsId = '') 
     drawerExtraOpen: false,
     drawerView: null,
     initialData: {} as Category,
+    unlinkModalOpen: false,
   }));
+  const { t } = useI18n();
 
   const fetchCategorySettings = async (categoryId: number): Promise<Category | null> => {
     if (cache.value[categoryId]) {
@@ -45,9 +47,43 @@ export const useCategorySettings: useCategorySettingsReturn = (settingsId = '') 
       state.value.loading = false;
     }
   };
-
+  const toggleDeleteModal = (value: boolean) => {
+    state.value.unlinkModalOpen = value;
+  };
+  const deletePage = async (id: number, pageName: string) => {
+    const { send } = useNotification();
+    try {
+      const { data } = await useSdk().plentysystems.deleteCategory({
+        categoryId: id,
+      });
+      if (data === 'Category deleted') {
+        toggleDeleteModal(false);
+        send({
+          message: t('errorMessages.editor.categories.deleteSuccess', { pageName: pageName, id: id }),
+          type: 'positive',
+        });
+      }
+    } catch (error) {
+      let errorMessage = '';
+      if (error && typeof error === 'object') {
+        if ('message' in error) {
+          errorMessage = (error as { message: string }).message;
+        }
+      }
+      send({
+        message: t('errorMessages.editor.categories.deleteError', {
+          pageName: pageName,
+          id: id,
+          errorMessage: errorMessage,
+        }),
+        type: 'negative',
+      });
+    }
+  };
   return {
     ...toRefs(state.value),
     fetchCategorySettings,
+    toggleDeleteModal,
+    deletePage,
   };
 };
