+++ apps/web/composables/useSiteConfiguration/useSiteConfiguration.ts
@@ -27,7 +27,6 @@ export const useSiteConfiguration: UseSiteConfigurationReturn = () => {
     pageModalOpen: false,
     settingsCategory: null,
     settingsType: null,
-    unlinkModalOpen: false,
     loading: false,
     placement: 'left',
     newBlockPosition: 0,
@@ -46,7 +45,6 @@ export const useSiteConfiguration: UseSiteConfigurationReturn = () => {
       secondaryColor: useRuntimeConfig().public.secondaryColor,
     },
   }));
-  const { t } = useI18n();
 
   /**
    * @description Function for loading a google font.
@@ -182,43 +180,9 @@ export const useSiteConfiguration: UseSiteConfigurationReturn = () => {
     state.value.pageModalOpen = value;
   };
 
-  const toggleDeleteModal = (value: boolean) => {
-    state.value.unlinkModalOpen = value;
-  };
   const setSettingsCategory = (category: CategoryTreeItem | null, settingsType?: SettingsType) => {
     state.value.settingsType = settingsType || null;
     state.value.settingsCategory = category;
-  };
-
-  const deletePage = async (id: number, pageName: string) => {
-    const { send } = useNotification();
-    try {
-      const { data } = await useSdk().plentysystems.deleteCategory({
-        categoryId: id,
-      });
-      if (data === 'Category deleted') {
-        toggleDeleteModal(false);
-        send({
-          message: t('errorMessages.editor.categories.deleteSuccess', { pageName: pageName, id: id }),
-          type: 'positive',
-        });
-      }
-    } catch (error) {
-      let errorMessage = '';
-      if (error && typeof error === 'object') {
-        if ('message' in error) {
-          errorMessage = (error as { message: string }).message;
-        }
-      }
-      send({
-        message: t('errorMessages.editor.categories.deleteError', {
-          pageName: pageName,
-          id: id,
-          errorMessage: errorMessage,
-        }),
-        type: 'negative',
-      });
-    }
   };
 
   return {
@@ -234,7 +198,5 @@ export const useSiteConfiguration: UseSiteConfigurationReturn = () => {
     saveSettings,
     togglePageModal,
     setSettingsCategory,
-    toggleDeleteModal,
-    deletePage,
   };
 };
